Responsiveness Fixes;
For screens without gesture navigation, the bottom padding was too much. This was because it was relying on a constant bottom nav bar height defined by flutter. This was replaced by a call to MediaQuery which fetches the bottom padding value at runtime. Bottom padding should render responsively based on available height now.

Usability on larger screens demands that actionable components should begin below a baseline within the reach of the average thumb. In About Us and List Routes screens, this baseline is implemented by passing a large top padding in the listview. This top padding was made responsive across small, medium and large screens.

Font size on Search Bar Suggestion chips was removed as the fixed size was too small.

UIs made for phones can have UI components that look too wife when rendered on a tablet. This was the case for footers across Go, List Routes and Create Route screens. A width constraint was added to make it look better.